Move DeviceError to utils
