+++ cheetah/utils.py
@@ -297,3 +297,16 @@ def load_cheetah_model(fname: str, name: Optional[str] = None) -> cheetah.Segmen
 def str_to_class(classname: str):
     # get class from string
     return getattr(cheetah, classname)
+
+
+class DeviceError(Exception):
+    """
+    Used to create an exception, in case the device used for the beam
+    and the elements are different.
+    """
+
+    def __init__(self):
+        super().__init__(
+            "Warning! The device used for calculating the elements is not the same, "
+            "as the device used to calculate the Beam."
+        )
